+++ src/iter.rs
@@ -347,3 +347,103 @@ where T: 'a + Node,
         self.index_iter.len()
     }
 }
+
+
+
+///////////////////////////
+/// * same as NContainedIter but mutable
+/// * Iterator over mutable additional information stored at vertices
+/// that are directly connected to specific vertex
+pub struct NContainedIterMut<'a, T, A>
+where T: Node,
+      A: AdjContainer<T>
+{
+    vertex_slice:   &'a mut [A],
+    index_iter:     IterWrapper<'a>,
+    phantom:        PhantomData<T>
+
+}
+
+impl<'a, T, A> NContainedIterMut<'a, T, A>
+where T: Node,
+      A: AdjContainer<T>
+{
+    /// Create new iterator
+    pub(crate) fn new(vertex_slice: &'a mut[A], index_iter: IterWrapper::<'a>) -> Self {
+        Self {
+            vertex_slice,
+            index_iter,
+            phantom: PhantomData::<T>
+        }
+    }
+}
+
+
+impl<'a, T, A> Iterator for NContainedIterMut<'a, T, A>
+where T: 'a + Node,
+      A: AdjContainer<T>
+{
+    type Item = &'a mut T;
+
+    #[inline]
+    fn next(&mut self) -> Option<Self::Item> {
+        let index = self.index_iter.next()?;
+        let index = *index as isize;
+
+        assert!(index < self.vertex_slice.len() as isize);
+
+        let ptr = self.vertex_slice.as_mut_ptr();
+        let r1: &mut A = unsafe { &mut *ptr.offset(index) };
+
+        Some(r1.contained_mut())
+    }
+
+    #[inline]
+    fn nth(&mut self, n: usize) -> Option<Self::Item> {
+        let index = self.index_iter.nth(n)?;
+        let index = *index as isize;
+
+        assert!(index < self.vertex_slice.len() as isize);
+
+        let ptr = self.vertex_slice.as_mut_ptr();
+        let r1: &mut A = unsafe { &mut *ptr.offset(index) };
+
+        Some(r1.contained_mut())
+    }
+
+    #[inline]
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        (self.len(), Some(self.len()))
+    }
+
+}
+
+impl<'a, T, A> DoubleEndedIterator for NContainedIterMut<'a, T, A>
+where T: 'a + Node,
+      A: AdjContainer<T>
+{
+    #[inline]
+    fn next_back(&mut self) -> Option<Self::Item> {
+        let index = self.index_iter.next_back()?;
+
+        let index = *index as isize;
+
+        assert!(index < self.vertex_slice.len() as isize);
+
+        let ptr = self.vertex_slice.as_mut_ptr();
+        let r1: &mut A = unsafe { &mut *ptr.offset(index) };
+
+        Some(r1.contained_mut())
+    }
+}
+
+impl<'a, T, A> ExactSizeIterator for NContainedIterMut<'a, T, A>
+where T: 'a + Node,
+      A: AdjContainer<T>
+{
+
+    #[inline]
+    fn len(&self) -> usize {
+        self.index_iter.len()
+    }
+}